Handle null cm_token consistently
ib_services now fails if the content of `cm_token` is null, and we have been asked to either send it if it's set, or not send it at all.  We already do this for other ib_services endpoints like A/PTR/CNAME, so making the behaviour in SRV consistent.

Addresses-Issue: AQUILONAQD-2969